3. Интерфейс Input [#320206]

+++ src/ru/job4j/poly/Bus.java
@@ -1,16 +1,18 @@
 package ru.job4j.poly;
 
-  public class Bus implements Transport{
+public class Bus implements Transport {
     @Override
     public void drive() {
+
     }
 
     @Override
     public void passengers(int amount) {
+
     }
 
     @Override
-    public int refuel(double fuel, double price) {
+    public int refuel(int fuel, double price) {
         return 0;
     }
 }
